Auth Commit
Returned register post method to the way it was before. Created a logout endpoint. Modified home page to be shown only when a user is logged in, and implemented a logout button. Added a condition to check if the user has already logged in for the login and register endpoints, if the user is logged, it will be redirected to the homePage endpoint. modified the password_hash field for the user table, this field was set as UNIQUE by mistake, this prevented different users to have the same password, as the hash generated would've been the same, breaking the UNIQUE condition